+++ src/main/java/seedu/address/logic/commands/EditCommand.java
@@ -20,7 +20,7 @@ import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
-import seedu.address.model.internship.*;
+import seedu.address.model.internshipapplication.*;
 import seedu.address.model.tag.Tag;
 
 /**
