+++ src/main/java/seedu/address/logic/parser/AddCommandParser.java
@@ -10,7 +10,7 @@ import java.util.stream.Stream;
 
 import seedu.address.logic.commands.AddCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.internship.*;
+import seedu.address.model.internshipapplication.*;
 
 /**
  * Parses input arguments and creates a new AddCommand object
