+++ src/main/java/seedu/address/logic/parser/FindCommandParser.java
@@ -6,7 +6,7 @@ import java.util.Arrays;
 
 import seedu.address.logic.commands.FindCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.internship.NameContainsKeywordsPredicate;
+import seedu.address.model.internshipapplication.NameContainsKeywordsPredicate;
 
 /**
  * Parses input arguments and creates a new FindCommand object
