+++ src/main/java/seedu/address/model/internshipapplication/exceptions/DuplicatePersonException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.internship.exceptions;
+package seedu.address.model.internshipapplication.exceptions;
 
 /**
  * Signals that the operation will result in duplicate Persons (Persons are considered duplicates if they have the same
