+++ src/main/java/seedu/address/model/internshipapplication/exceptions/InternshipNotFoundException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.internship.exceptions;
+package seedu.address.model.internshipapplication.exceptions;
 
 /**
  * Signals that the operation is unable to find the specified internship.
