+++ src/main/java/seedu/address/model/util/SampleDataUtil.java
@@ -7,7 +7,7 @@ import java.util.stream.Collectors;
 
 import seedu.address.model.AddressBook;
 import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.internship.*;
+import seedu.address.model.internshipapplication.*;
 import seedu.address.model.tag.Tag;
 
 /**
