+++ src/main/java/seedu/address/storage/JsonAdaptedInternship.java
@@ -7,7 +7,7 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.model.internship.*;
+import seedu.address.model.internshipapplication.*;
 
 /**
  * Jackson-friendly version of {@link Person}.
