+++ src/main/java/seedu/address/ui/InternshipCard.java
@@ -4,7 +4,7 @@ import javafx.fxml.FXML;
 import javafx.scene.control.Label;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
-import seedu.address.model.internship.InternshipApplication;
+import seedu.address.model.internshipapplication.InternshipApplication;
 
 /**
  * An UI component that displays information of a {@code Person}.
