+++ src/test/java/seedu/address/logic/commands/AddCommandIntegrationTest.java
@@ -11,7 +11,7 @@ import seedu.address.logic.Messages;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.internship.Person;
+import seedu.address.model.internshipapplication.Person;
 import seedu.address.testutil.PersonBuilder;
 
 /**
