+++ src/test/java/seedu/address/logic/commands/AddCommandTest.java
@@ -22,7 +22,7 @@ import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.ReadOnlyUserPrefs;
-import seedu.address.model.internship.Person;
+import seedu.address.model.internshipapplication.Person;
 import seedu.address.testutil.PersonBuilder;
 
 public class AddCommandTest {
