+++ src/test/java/seedu/address/logic/parser/AddressBookParserTest.java
@@ -23,8 +23,8 @@ import seedu.address.logic.commands.FindCommand;
 import seedu.address.logic.commands.HelpCommand;
 import seedu.address.logic.commands.ListCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.internship.NameContainsKeywordsPredicate;
-import seedu.address.model.internship.Person;
+import seedu.address.model.internshipapplication.NameContainsKeywordsPredicate;
+import seedu.address.model.internshipapplication.Person;
 import seedu.address.testutil.EditPersonDescriptorBuilder;
 import seedu.address.testutil.PersonBuilder;
 import seedu.address.testutil.PersonUtil;
