+++ src/test/java/seedu/address/logic/parser/FindCommandParserTest.java
@@ -9,7 +9,7 @@ import java.util.Arrays;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.FindCommand;
-import seedu.address.model.internship.NameContainsKeywordsPredicate;
+import seedu.address.model.internshipapplication.NameContainsKeywordsPredicate;
 
 public class FindCommandParserTest {
