+++ src/test/java/seedu/address/testutil/TypicalPersons.java
@@ -16,7 +16,7 @@ import java.util.Arrays;
 import java.util.List;
 
 import seedu.address.model.AddressBook;
-import seedu.address.model.internship.Person;
+import seedu.address.model.internshipapplication.Person;
 
 /**
  * A utility class containing a list of {@code Person} objects to be used in tests.
